backend/requirements.txt
fastapi
uvicorn
paho-mqtt
roslibpy
prisma
